housekeeping coord2idx functions for grids

+++ src/grids/spherical2d.jl
@@ -205,37 +205,37 @@ end
 ############################################################################################
 #::. UTILITY FUNCTIONS
 ############################################################################################
-function coord2idx(grid::Spherical2DGrid, theta::Real, phi::Real)
-    theta = pclamp(theta, -pi + eps(Float64), pi)
-    phi = clamp(phi, -pi/2 + eps(Float64), pi/2)
+function coord2idx(grid::Spherical2DGrid, theta::Real, phi::Real)::Int64
+    theta    = pclamp(theta, -pi + eps(Float64), pi)
+    phi      = clamp(phi, -pi/2 + eps(Float64), pi/2)
     idxtheta = ceil(Int64, (theta+pi)*grid.N_theta/2/pi)
-    idxphi = ceil(Int64, (phi+pi/2)*grid.N_phi/pi)
+    idxphi   = ceil(Int64, (phi+pi/2)*grid.N_phi/pi)
     return (idxtheta-1) * grid.N_phi + idxphi
 end
-function coord2idx(grid::Spherical2DGrid_EqSim, theta::Real, phi::Real)
-    theta = pclamp(theta, -pi + eps(Float64), pi)
-    phi = clamp(phi, eps(Float64), pi/2)
+function coord2idx(grid::Spherical2DGrid_EqSim, theta::Real, phi::Real)::Int64
+    theta    = pclamp(theta, -pi + eps(Float64), pi)
+    phi      = clamp(phi, eps(Float64), pi/2)
     idxtheta = ceil(Int64, (theta+pi)*grid.N_theta/2/pi)
-    idxphi = ceil(Int64, abs(phi)*grid.N_phi*2/pi)
+    idxphi   = ceil(Int64, abs(phi)*grid.N_phi*2/pi)
     return (idxtheta-1) * grid.N_phi + idxphi
 end
-function coord2idx(grid::Spherical2DGrid_Reduced, theta::Real, phi::Real)
-    theta = pclamp(theta, -pi + eps(Float64), pi)
-    phi = clamp(phi, -pi/2 + eps(Float64), pi/2)
-    idxphi = ceil(Int64, (phi+pi/2)/pi*grid.N_phi)
+function coord2idx(grid::Spherical2DGrid_Reduced, theta::Real, phi::Real)::Int64
+    theta    = pclamp(theta, -pi + eps(Float64), pi)
+    phi      = clamp(phi, -pi/2 + eps(Float64), pi/2)
+    idxphi   = ceil(Int64, (phi+pi/2)/pi*grid.N_phi)
     idxtheta = ceil(Int64, (theta+pi)/2/pi*grid.N_theta[idxphi])
     if idxphi == 1; return idxtheta; end
     return idxtheta + accumulate(+, grid.N_theta[1:idxphi])[end-1]
 end
-function coord2idx(grid::Spherical2DGrid_Reduced_EqSim, theta::Real, phi::Real)
-    theta = pclamp(theta, -pi + eps(Float64), pi)
-    phi = clamp(phi, eps(Float64), pi/2)
-    idxphi = ceil(Int64, abs(phi)*grid.N_phi*2/pi)
+function coord2idx(grid::Spherical2DGrid_Reduced_EqSim, theta::Real, phi::Real)::Int64
+    theta    = pclamp(theta, -pi + eps(Float64), pi)
+    phi      = clamp(phi, eps(Float64), pi/2)
+    idxphi   = ceil(Int64, abs(phi)*grid.N_phi*2/pi)
     idxtheta = ceil(Int64, (theta+pi)*grid.N_theta[idxphi]/2/pi)
     if idxphi == 1; return idxtheta; end
     return idxtheta + accumulate(+, grid.N_theta[1:idxphi])[end-1]
 end
-function coord2idx(grid::AbstractSpherical2DGrid, r::Real, theta::Real, phi::Real)
+function coord2idx(grid::AbstractSpherical2DGrid, r::Real, theta::Real, phi::Real)::Int64
     return coord2idx(grid, theta, phi)
 end
